backend/requirements.txt
# Core FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Payment
stripe==7.11.0

# Email
aiosmtplib==3.0.1
sendgrid==6.11.0

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9

# Monitoring
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.39.2

# Core Utils
requests==2.31.0
httpx==0.26.0
pillow==10.2.0
reportlab==4.0.8
twilio==8.10.0
psutil==5.9.6

# AI/ML - Lightweight versions only
openai==1.6.1
numpy==1.26.3